+++ module/CLI/src/Input/VisitsListInput.php
@@ -10,7 +10,7 @@ use Symfony\Component\Console\Attribute\Option;
 use function Shlinkio\Shlink\Common\buildDateRange;
 use function Shlinkio\Shlink\Core\normalizeOptionalDate;
 
-class VisitsDateRangeInput
+class VisitsListInput
 {
     #[Option('Only return visits older than this date', shortcut: 's')]
     public string|null $startDate = null;
@@ -18,7 +18,14 @@ class VisitsDateRangeInput
     #[Option('Only return visits newer than this date', shortcut: 'e')]
     public string|null $endDate = null;
 
-    public function toDateRange(): DateRange
+    #[Option(
+        'Output format ("' . VisitsListFormat::FULL->value . '", "' . VisitsListFormat::PAGINATED->value . '" or "'
+        . VisitsListFormat::CSV->value . '")',
+        shortcut: 'f',
+    )]
+    public VisitsListFormat $format = VisitsListFormat::FULL;
+
+    public function dateRange(): DateRange
     {
         return buildDateRange(
             startDate: normalizeOptionalDate($this->startDate),